seo_analyzer/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
python-dotenv==1.0.0